feat(AIButton): update stories for AI Button component
