+++ core/ai-components/AIButton/__stories__/appearance/primary.story.jsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import { AIButton } from '@/index';
 
-export const Basic = () => {
-  return <AIButton>Basic Button</AIButton>;
+export const Primary = () => {
+  return <AIButton appearance="primary">Primary Button</AIButton>;
 };
 
 export default {
-  title: 'Components/AI/AI Button/States/Basic',
+  title: 'Components/AI/AI Button/Appearance/Primary',
   component: AIButton,
   parameters: {
     docs: {
