+++ core/ai-components/AIButton/__stories__/states/basic.story.jsx
@@ -1,17 +1,17 @@
 import React from 'react';
 import { AIButton } from '@/index';
 
-export const Appearance = () => {
+export const Basic = () => {
   return (
     <div className="d-flex justify-content-between w-50">
-      <AIButton appearance="primary">Primary Button</AIButton>
       <AIButton>Basic Button</AIButton>
+      <AIButton disabled={true}>Disabled Button</AIButton>
     </div>
   );
 };
 
 export default {
-  title: 'Components/AI/AI Button/Appearance',
+  title: 'Components/AI/AI Button/State/Basic',
   component: AIButton,
   parameters: {
     docs: {
