+++ core/index.type.tsx
@@ -94,3 +94,4 @@ export { SelectProps } from './components/organisms/select';
 export { MenuProps } from './components/organisms/menu';
 
 export { SaraSparkleProps } from './ai-components/SaraSparkle';
+export { AIButtonProps } from './ai-components/AIButton';
