+++ css/src/ai-components/button.css
@@ -56,6 +56,7 @@
 
 .AIButton--basic:disabled {
   background: var(--secondary-lighter);
+  color: var(--inverse-lightest);
 }
 
 .AIButton--basic:focus {
